+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.4.21 (Unreleased)
 
 IMPROVEMENTS:
+* provider: add `skip_verify` configuration to skip TLS verification ([#319](https://github.com/hashicorp/terraform-provider-nomad/pull/319))
 * resource/nomad_acl_policy: add support for `job_acl` ([#314](https://github.com/hashicorp/terraform-provider-nomad/pull/314))
 
 BUG FIXES: